refactor: 应用 clippy 建议并重组 DPI 感知模块
- 将 set_thread_dpi_awareness_context 从 lib.rs 移至 window 模块
- screencap_region 参数类型 Region2D<u32> 改为 Region2D<i32>
- unsafe 块范围缩小、MSG::default() 替代 mem::zeroed()
- abs_diff 替代手动绝对值、entry API 替代 contains_key+insert
- 移除冗余枚举索引、as 类型转换，使用 Self:: 消歧义 trait 调用

+++ src/screencap/base.rs
@@ -5,18 +5,20 @@ use windows::Win32::Foundation::HWND;
 use crate::utils::region::Region2D;
 
 pub trait ScreencapBase {
-    fn new(hwnd: HWND) -> Self where Self: Sized;
+    fn new(hwnd: HWND) -> Self
+    where
+        Self: Sized;
 
     fn screencap(&mut self) -> Result<RgbaImage>;
 
-    fn screencap_region(&mut self, relative_region: Region2D<u32>) -> Result<RgbaImage> {
+    fn screencap_region(&mut self, relative_region: Region2D<i32>) -> Result<RgbaImage> {
         let image = self.screencap()?;
         let cropped = imageops::crop_imm(
             &image,
-            relative_region.x0(),
-            relative_region.y0(),
-            relative_region.width(),
-            relative_region.height(),
+            relative_region.x0() as u32,
+            relative_region.y0() as u32,
+            relative_region.width() as u32,
+            relative_region.height() as u32,
         )
         .to_image();
         Ok(cropped)
